feature: accepting optional container in docker:up, flags for tine:test

+++ cli/Commands/Tine/TineInstallCommand.php
@@ -6,6 +6,7 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use App\ConsoleStyle;
 use App\Commands\Tine\TineCommand;
@@ -17,22 +18,32 @@ class TineInstallCommand extends TineCommand{
             ->setName('tine:install')
             ->setDescription('install tine')
             ->setHelp('')
-        ;
+            ->addArgument(
+                'modules',
+                InputArgument::OPTIONAL | InputArgument::IS_ARRAY,
+                'The modules you want to install'
+            );
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $io = new ConsoleStyle($input, $output);
-        
+        $inputOptions = $input->getArgument('modules');
         $this->initCompose();
-        if ($this->active('mailstack')) {
-            $this->mailstackInit($io);
-            $this->mailstackReset($io);
+        
+        if(empty($inputOptions)) {
+            if ($this->active('mailstack')) {
+                $this->mailstackInit($io);
+                $this->mailstackReset($io);
+            }
+    
+            passthru($this->getComposeString() . ' exec -T cache sh -c "redis-cli flushall"', $err);
+            $io->notice("Installing Tine 2.0 ...");
+            passthru($this->getComposeString() . ' exec -T --user tine20 web tine20_install', $err);
+        }else {
+            passthru($this->getComposeString() . ' exec -T --user tine20 web sh -c "cd tine20 && php setup.php --install "'
+            . implode(" ", $inputOptions), $err);
         }
-
-        passthru($this->getComposeString() . ' exec -T cache sh -c "redis-cli flushall"', $err);
-        $io->notice("Installing Tine 2.0 ...");
-        passthru($this->getComposeString() . ' exec -T --user tine20 web tine20_install', $err);
 
         return Command::SUCCESS;
     }
